[ENG-831] Add framework to support system column id for row level TTL.
Summary:
Few changes in this diff
1) Change columnid from int64 to uint32_t. (Discussed offline with akashnil that 2^32 columns should be sufficient)
2) Enhance read and write path to accomodate new column id.
3) Introduce system columns to support the row level TTL usecase.
3) Use uint8_t to support system column ids (256 should be sufficient).

Test Plan: unit tests, jenkins

Reviewers: hector, akashnil, mikhail

Reviewed By: mikhail

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D1189

+++ src/yb/docdb/doc_rowwise_iterator.cc
@@ -210,7 +210,7 @@ Status DocRowwiseIterator::GetValues(const Schema& projection, vector<PrimitiveV
     const auto& column_id = projection.column_id(i);
     subdoc_key_.RemoveHybridTime();
     subdoc_key_.AppendSubKeysAndMaybeHybridTime(
-        PrimitiveValue(static_cast<int32_t>(column_id)), hybrid_time_);
+        PrimitiveValue(column_id), hybrid_time_);
 
     const KeyBytes key_for_column = subdoc_key_.Encode();
     ROCKSDB_SEEK(db_iter_.get(), key_for_column.AsSlice());
